pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "AI-powered recruitment system with FastAPI backend"
authors = [
    {name = "Abdelmoneim Mohamed",email = "96942944+Xmen3em@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "phidata (==2.7.3)",
    "pypdf2 (==3.0.1)",
    "requests (==2.32.3)",
    "pytz (==2023.4)",
    "typing-extensions (>=4.9.0)",
    "black (>=24.1.1)",
    "python-dateutil (>=2.8.2)",
    "fastapi (>=0.115.0)",
    "uvicorn[standard] (>=0.32.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.6.0)",
    "python-multipart (>=0.0.20)",
    "openai (>=2.8.1,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "agent", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
